Move key mappings to the top of the script (#11)
* Move "I already ran" check so key mappings are always done.

This plugin has a check in it to prevent the recreation of functions
while they are being executed. (See issue #6.) The placement of this
check at the very top of the script prevents the creation of the key
mappings on all but the first NERDTree buffer. Moving the check to just
after the key mappings section and before all the function definitions
solves this issue.