Allow cancel of Move/Copy by pressing Esc.
Before the script loops through the selection, it will ask the user for
the path into which the files are to be moved or coppied. If the user
presses Esc now when asked for the destination, the process will bypass
the loop altogether, and no copying or moving will take place.